refactor: colocate recurring quest generator tests
