+++ lib/recurring-quest-generator.fixtures.ts
@@ -1,5 +1,5 @@
 import type { SupabaseClient } from "@supabase/supabase-js";
-import type { Database } from "../types/database-generated";
+import type { Database } from "./types/database-generated";
 
 export const createMockSupabase = () =>
   ({
